Move builtin string_insert constraint code to new file
Put all the code related to the builtin string_insert function to a
string_insertion_builtin_function file.
